♻️ refactor: rename useExtensionState to useQuickMenuState
