--- src/presentation/hooks/useExtensionState/index.ts
@@ -1,1 +1,0 @@
-export { useExtensionState } from './useExtensionState';
